refactor(ingester): Split up the `WriteOperation` sub-types into separate modules
