+++ ingester/src/dml_payload/write/mod.rs
@@ -1,0 +1,9 @@
+//! This submodule holds types constructed from information in a write
+mod op;
+pub use op::*;
+
+mod partitioned_data;
+pub use partitioned_data::*;
+
+mod table_data;
+pub use table_data::*;
